Move territoires.js in /server
