+++ server/projets-validator.js
@@ -1,8 +1,8 @@
 /* eslint-disable camelcase */
 
 import Joi from 'joi'
-import {validatePayload} from '../server/util/payload.js'
-import {ensureTerritoireExists} from '../lib/territoires.js'
+import {validatePayload} from './util/payload.js'
+import {ensureTerritoireExists} from './territoires.js'
 
 function validatePerimetre(perimetre) {
   const fields = perimetre.split(':')
