+++ server/projets.js
@@ -1,8 +1,8 @@
 import createError from 'http-errors'
 import {omit} from 'lodash-es'
 import {nanoid} from 'nanoid'
-import {buildGeometryFromTerritoires, getTerritoiresProperties} from '../lib/territoires.js'
 import {findClosestEtape} from '../shared/find-closest-etape.js'
+import {buildGeometryFromTerritoires, getTerritoiresProperties} from './territoires.js'
 import {validateCreation, validateChanges} from './projets-validator.js'
 import mongo from './util/mongo.js'
